+++ docs/changes.rst
@@ -6,6 +6,8 @@ weblate 2.7
 
 Released on ? 2016.
 
+* Removed Google web translate machine translation.
+
 weblate 2.6
 -----------